[MAINT] Reorganize GLM benchmarks to mirror nilearn.glm's own structure
save_glm_to_bids is permanently moving to nilearn.glm.io in nilearn
0.15.0; nilearn.interfaces.bids.save_glm_to_bids (its current,
deprecated, location) will be removed then. Move
BenchMarkFirstLevelModelSave out of glm.py into a new
asv_benchmarks/benchmarks/glm/io.py, mirroring nilearn/glm/io.py, and
import save_glm_to_bids locally in setup() with a fallback: try
nilearn.interfaces.bids first, then nilearn.glm.io if that import
fails, following the local-import pattern documented in
CONTRIBUTING.rst so this only fails this one benchmark rather than the
whole module for nilearn versions on either side of the move.

Also move BenchMarkFirstLevelModel and BenchMarkFirstLevelModelReport
out of glm.py into asv_benchmarks/benchmarks/glm/first_level/first_level.py,
mirroring the nilearn/glm/first_level/ subpackage. glm.py now only
holds the shared BaseBenchMarkFLM base class.

Document in CONTRIBUTING.rst that the asv_benchmarks/benchmarks
directory structure must mirror nilearn's own package structure, and
that only public API functions/methods should be benchmarked.

Verified locally that all three benchmark classes are discovered and
run correctly from their new locations, including the
save_glm_to_bids fallback resolving nilearn.interfaces.bids first (as
expected pre-0.15.0, with its deprecation FutureWarning visible).


[skip doc] [skip test]
